New NewId algorithm and some minor changes (#247)
+ More reading and using DnsSuffix
+ Change default NetworkInterface discovery to remove unknown interfaces and order by type

+++ src/DnsClient/DnsMessageHandler.cs
@@ -125,7 +125,7 @@ namespace DnsClient
                 writer.WriteHostName("");
                 writer.WriteUInt16NetworkOrder((ushort)opt.RecordType);
                 writer.WriteUInt16NetworkOrder((ushort)opt.RecordClass);
-                writer.WriteUInt32NetworkOrder((ushort)opt.InitialTimeToLive);
+                writer.WriteUInt32NetworkOrder((uint)opt.InitialTimeToLive);
                 writer.WriteUInt16NetworkOrder(0);
             }
         }
